backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
numpy>=1.26.0
pydantic>=2.0.0
openai>=1.12.0
requests==2.31.0
langchain>=0.2.0
langchain-community>=0.0.10
langchain-upstage
tzdata==2023.3
python-multipart>=0.0.7
typing-extensions>=4.5.0
aiohttp>=3.8.0
spacy>=3.0.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
backports.zoneinfo;python_version<"3.9"  
pytz>=2024.1
starlette>=0.36.0